Cleans up App.js and adds shutdown option to the tool bar
