+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Adds shutdown option to the tool bar
+
 ## [0.1.12] - 2018-01-09
 
 ### Changes
