+++ src/components/views/LogView.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function LogView ({ messages }) {
+    messages = messages || [];
+    return (
+        <pre className='logview'>
+            { messages.join('\n') }
+        </pre>
+    );
+}
